backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.7.4,<3.0.0
python-multipart==0.0.6

# Azure Services
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-cosmos==4.5.1

# Computer Vision & ML
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2

# Pose Estimation
mmpose==1.1.0
mmcv==2.1.0
mmengine==0.10.0

# 3D Processing
smplx==0.1.28
trimesh==3.23.5
pyrender==0.1.45

# Temporal Processing
transformers==4.35.0
einops==0.7.0

# Signal Processing
filterpy==1.4.5
scipy==1.11.4

# Data Processing
pandas==2.1.3
h5py==3.10.0

# Medical Data Formats
fhir.resources==8.0.0

# Utilities
python-dotenv==1.0.0
pydantic-settings==2.1.0
loguru==0.7.2
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

